fix: LockerId Integer로 변경-Repository/#2

+++ src/main/java/locker/repository/LockerRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface LockerRepository {
 
-    Optional<Locker> getLocker(Long id);
+    Optional<Locker> getLocker(Integer id);
     void replaceLocker(Locker locker);
     List<Locker> getLockersInUse();
     boolean isAllInUse();
